+++ CHANGELOG.md
@@ -14,4 +14,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Removed
 
-## [M.m.p] - YYYY-MM-DD
+## [0.1.0] - 2025-08-06
+
+### Added
+
+- Birth Number calculation
+- Personal Day calculation
+
+### Changed
+
+- Rewrote functions for more simplified calculation (all previous functions improved)
+- Updated menu options
+    - Add Birth Chart option
+    - Add Comparison Chart option
+    - Add Prediction Chart option
+
+## [0.0.0] - 2024-08-05
+
+### Added
+
+- Universal Year calculation
+- Universal Month calculation
+- Personal Year calculation
+- Personal Month calculation (not fully functional)